+++ src/world/physics.h
@@ -9,6 +9,14 @@
 
 namespace crown
 {
+/// Global physics settings.
+/// @ingroup World
+struct PhysicsSettings
+{
+	s32 step_frequency;
+	s32 max_substeps;
+};
+
 /// Global physics-related functions
 ///
 /// @ingroup World
@@ -16,7 +24,7 @@ namespace physics_globals
 {
 	/// Initializes the physics system.
 	/// This is the place where to create and initialize per-application objects.
-	void init(Allocator &a);
+	void init(Allocator &a, const PhysicsSettings *settings);
 
 	/// It should reverse the actions performed by physics_globals::init().
 	void shutdown(Allocator &a);